Refactor branding classes, improve color parsing, bugfixes
- Merged SpaBrandingManifest, SpaBrandingColors, and SpaBrandingIcons into SpaBrandingManifest.cs as internal sealed classes in VpnHood.AppLib.SpaWebView.
- Updated SpaResourcesFactory to use new manifest location and VhColor.TryParse for safer color parsing.
- Added VhColor.TryParse; enhanced Parse error messages.
- Simplified VhTypeTracker references.
- Minor code style and formatting improvements.
- Fixed poll() argument type in LinuxTunVpnAdapter.
- Improved IPv6 split-tunnel logic in IosVpnAdapter.

+++ src/Core/VpnHood.Core.VpnAdapters.LinuxTun/LinuxTunVpnAdapter.cs
@@ -281,7 +281,7 @@ public class LinuxTunVpnAdapter(LinuxVpnAdapterSettings adapterSettings)
     private static void WaitForTun(StructPollfd[] pollFds)
     {
         while (true) {
-            var result = LinuxAPI.poll(pollFds, (nuint)1, -1);
+            var result = LinuxAPI.poll(pollFds, 1, -1);
             if (result >= 0)
                 break; // Success, exit loop
